requirements.txt
alembic==1.7.3
angles==2.0
asgiref==3.4.1
boto==2.49.0
botocore==1.16.5
certifi==2020.4.5.1
chardet==3.0.4
click==8.0.1
dbbackup==0.1.13
dbbackup-tools==0.1.5
Django==3.2.5
django-dbbackup==3.2.0
django-filter==21.1
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.12.4
docutils==0.15.2
ecdsa==0.17.0
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.1.0
idna==2.10
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==2.11.2
jmespath==0.10.0
jose==1.0.0
Mako==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.21.1
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
pkg_resources==0.0.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pypng==0.0.20
python-dateutil==2.8.2
python-decouple==3.4
python-editor==1.0.4
python-jose==3.1.0
pytz==2021.1
PyYAML==5.4.1
pyzbar==0.1.8
qrcode==7.2
random2==1.0.1
requests==2.23.0
rsa==4.0
secure-smtplib==0.1.1
six==1.16.0
SQLAlchemy==1.3.16
sqlparse==0.4.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.3.0
xacro==1.13.3
XlsxWriter==3.0.1
zipp==3.6.0
